% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureApproximation.R
\name{approximateFeatures}
\alias{approximateFeatures}
\title{Compute approximate features used by the OKRF tree in splitting}
\usage{
approximateFeatures(feat_rep, tol)
}
\description{
Compute approximate features used by the OKRF tree in splitting
}
