% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureApproximation.R
\name{pqr_wrap}
\alias{pqr_wrap}
\title{Compute approximate features by pivoted QR used in splitting}
\usage{
pqr_wrap(X, tol)
}
\description{
Compute approximate features by pivoted QR used in splitting
}
