% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureApproximation.R
\name{prepareFeatures}
\alias{prepareFeatures}
\title{Prepare features used by the OKRF tree}
\usage{
prepareFeatures(feat_rep, tol, scope = "forest", reference_ids = NULL)
}
\description{
`feat_rep` contains the original output representation, either `Phi`
or `K`. The returned object separates features used for splitting
from features used for terminal response predictions.
}
