% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleOKRF.R
\name{simpleOKRF}
\alias{simpleOKRF}
\title{simpleOKRF}
\usage{
simpleOKRF(
  K = NULL,
  Phi = NULL,
  X,
  tol = 0.001,
  num_trees = 200,
  mtry = NULL,
  min_node_size = NULL,
  replace = TRUE,
  unordered_factors = "ignore",
  num_threads = 1
)
}
\arguments{
\item{K}{Gram matrix of targets. If possible, use Phi instead of K.}

\item{Phi}{Feature matrix of targets (\eqn{n \times d, n > d}).}

\item{X}{Covariate data of class \code{data.frame}, with one row per target observation.}

\item{tol}{Error tolerance of approximation. Default 0.001.}

\item{num_trees}{Number of trees.}

\item{mtry}{Number of variables to possibly split at in each node.}

\item{min_node_size}{Minimal node size. Default 5.}

\item{replace}{Sample with replacement. Default TRUE.}

\item{unordered_factors}{How to handle unordered factor variables. Either "ignore" or "partition" with default "ignore".}

\item{num_threads}{Number of threads used for mclapply, set to 1 for debugging.}
}
\description{
A simple implementation of Output Kernel Random Forests using forest-wise
feature approximations based on the R package \code{simpleRF}.
Uses reference classes and only plain \code{R}.
Not optimized for computation speed.
}
\details{
Unordered factor variables can be handled in different ways.
Use "ignore" to treat them as ordered in the order of the factor levels.
With "partition" all 2-partitions of the factor levels are considered for splitting.
}
\examples{
\donttest{
set.seed(20260916)

## Covariates
n <- 150

X <- data.frame(x1 = runif(n, -3, 3), x2 = runif(n, -3, 3))

## Scalar output signal
y <- sin(X$x1) + 0.5 * cos(2 * X$x2) + rnorm(n)

## Random Fourier Features for a Gaussian output kernel
num_frequencies <- 40
ell <- 1

omega <- rnorm(num_frequencies, mean = 0, sd = 1 / ell)

projection <- outer(y, omega)

Phi <- sqrt(1 / num_frequencies) * cbind(cos(projection), sin(projection))

## Fit an Output Kernel Random Forest
forest <- simpleOKRF(
  Phi = Phi,
  X = X,
  tol = 1e-6,
  num_trees = 25,
  mtry = 1,
  min_node_size = 5,
  replace = TRUE,
  unordered_factors = "ignore",
  num_threads = 1
)

## Predict the output representation for new covariates
newdata <- data.frame( x1 = c(-2, 0, 2), x2 = c(-2, 0, 2))

Phi_prediction <- forest$predict(
  newdata = newdata,
  type = "response"
)

dim(Phi_prediction)
}
}
\author{
Luca Bergen
}
